% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.r
\name{lm_fit}
\alias{lm_fit}
\title{lm_fit}
\usage{
lm_fit(x, y)
}
\arguments{
\item{x}{Feature matrix.}

\item{y}{Response variable.}
}
\value{
A list containing the coefficients, residuals, and effects of the model.
}
\description{
Basic linear model fitter.
}
