% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQ.r
\name{LQ}
\alias{LQ}
\title{LQ}
\usage{
LQ(x, retl = TRUE, retq = TRUE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{retl, retq}{Should L and/or Q be returned?}
}
\value{
If both \code{retq} and \code{retr}, then a list containing \code{Q} and
\code{R} is returned.  Otherwise, just the requested object is returned.
If neither \code{retq} and \code{retr} is \code{TRUE}, then the function
errors.
}
\description{
Factors a matrix x = LQ, where Q is orthogonal and L is lower triangular.
}
\details{
This is cheaper to compute than LQ when m<<n.

The LQ is computed via \code{dgelqf()}.  Q is extracted via \code{dorglq()}.
}
\examples{
\dontrun{
library(lqr)
m <- 3
n <- 10
x <- matrix(rnorm(m*n), m, n)

L <- LQ(x, retq=FALSE)

Q <- LQ(x, retl=FALSE)

LandQ <- LQ(x)

# error!
neither <- LQ(x, retl=FALSE, retq=FALSE)
}

}

