% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.r
\name{QR}
\alias{QR}
\title{QR}
\usage{
QR(x, retq = TRUE, retr = TRUE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{retq, retr}{Should Q and/or R be returned?}
}
\value{
If both \code{retq} and \code{retr}, then a list containing \code{Q} and
\code{R} is returned.  Otherwise, just the requested object is returned.
If neither \code{retq} and \code{retr} is \code{TRUE}, then the function
errors.
}
\description{
Factors a matrix x = QR, where Q is orthogonal and R is upper triangular.
}
\details{
This is cheaper to compute than LQ when m>>n.

The QR is computed via \code{dgeqrf()}.  Q is extracted via \code{dorgqr()}.
}
\examples{
\dontrun{
library(lqr)
m <- 10
n <- 3
x <- matrix(rnorm(m*n), m, n)

Q <- QR(x, retr=FALSE)

R <- QR(x, retq=FALSE)

QandR <- QR(x)

# error!
neither <- QR(x, retq=FALSE, retr=FALSE)
}

}

