% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQ.r
\name{LQ}
\alias{LQ}
\title{LQ}
\usage{
LQ(x, retl = TRUE, retq = TRUE)
}
\arguments{
\item{x}{}

\item{retl, retq}{Should L and/or Q be returned?}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
library(lqr)
m <- 3
n <- 10
x <- matrix(rnorm(m*n), m, n)

L <- LQ(x, retq=FALSE)

Q <- LQ(x, retl=FALSE)

LandQ <- LQ(x)

# error!
neither <- LQ(x, retl=FALSE, retq=FALSE)
}

}

