% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqr-package.r
\docType{package}
\name{lqr-package}
\alias{lqr-package}
\title{lqr:}
\description{
A streamlined interface for carefully optimized, highly high-performance
access to the QR and LQ factorizations.
}
\author{
Drew Schmidt
}
\keyword{package}

