% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.r
\name{QR}
\alias{QR}
\title{QR}
\usage{
QR(x, retq = TRUE, retr = TRUE)
}
\arguments{
\item{x}{}

\item{retq, retr}{Should Q and/or R be returned?}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
library(lqr)
m <- 10
n <- 3
x <- matrix(rnorm(m*n), m, n)

Q <- QR(x, retr=FALSE)

R <- QR(x, retq=FALSE)

QandR <- QR(x)

# error!
neither <- QR(x, retq=FALSE, retr=FALSE)
}

}

